refactor: remove v2 routes directory
