+++ test/integration/siti-agri/agreements.test.js
@@ -1,6 +1,6 @@
 import Hapi from '@hapi/hapi'
 import { retrieveOrganisationAgreements } from '../../../src/factories/siti-agri/agreement.factory.js'
-import { sitiagri } from '../../../src/routes/v2/siti-agri.js'
+import { sitiagri } from '../../../src/routes/siti-agri.js'
 import { loadSchema } from '../../../src/utils/validatePayload.js'
 
 describe('Basic queries for faked routes', () => {
@@ -10,7 +10,7 @@ describe('Basic queries for faked routes', () => {
     server.route(sitiagri)
     await Promise.all([
       server.initialize(),
-      loadSchema('routes/v2/siti-agri-schema.oas.yml').then((s) => (schema = s))
+      loadSchema('routes/siti-agri-schema.oas.yml').then((s) => (schema = s))
     ])
   })
